Update HSbConverter.cs
As I recently found out myself, HSB is NOT equivalent to HSL. It's actually equivalent to HSV, so I changed this specific code. (I have not tested it, just followed the existing pattern.)

It's a common mistake to confuse HSB and HSL, even by Microsoft in System.Drawing.Color, where MSDN plainly states that GetBrightness() and GetSaturation() are returning HSB values, when in fact they're returning HSL. The difference between HSL and HSB is apparent with something like red (#ff0000). HSB/HSV would be (0, 1, 1), while HSL would be (0, 1, .5).